refactor: 계산 설명 화면 SingleButtonScreen으로 변경
